Refactoring; new boolean expression code moved to synthesis, old code restored.
