+++ qiskit/circuit/library/__init__.py
@@ -345,6 +345,7 @@ quantum circuits of interest:
    QuantumVolume
    PhaseEstimation
    GroverOperator
+   BitFlipOracle
    PhaseOracle
    PauliEvolutionGate
    HamiltonianGate
@@ -671,5 +672,6 @@ from .iqp import IQP, iqp, random_iqp
 from .phase_estimation import PhaseEstimation, phase_estimation
 from .grover_operator import GroverOperator, grover_operator
 from .phase_oracle import PhaseOracle
+from .bit_flip_oracle import BitFlipOracle
 from .overlap import UnitaryOverlap, unitary_overlap
 from .standard_gates import get_standard_gate_name_mapping
