+++ releasenotes/notes/boolean_expression_update-39c19edbbe71ba0d.yaml
@@ -1,6 +1,8 @@
 ---
 upgrade_circuits:
   - |
-    The :class:`.BooleanExpression` no longer relies on the `tweedledum` library
+    The :class:`.PhaseOracle` no longer relies on the `tweedledum` library
     but might not be synthesize as effectively as before.
-    Passing custom synthesizers is no longer available.
+    :class:`.BitFlipOracle` was added as an alternative to
+    directly synthesizing :class:`.BooleanExpression`, as this
+    class will be removed in Qiskit 2.0.
