+++ test/python/classical_function_compiler/test_boolean_expression.py
@@ -18,10 +18,14 @@ from ddt import ddt, unpack, data
 
 from qiskit import transpile
 from qiskit.providers.basic_provider import BasicSimulator
+from qiskit.utils.optionals import HAS_TWEEDLEDUM
 from test import QiskitTestCase  # pylint: disable=wrong-import-order
-from qiskit.circuit.classicalfunction.boolean_expression import BooleanExpression
+
+if HAS_TWEEDLEDUM:
+    from qiskit.circuit.classicalfunction.boolean_expression import BooleanExpression
 
 
+@unittest.skipUnless(HAS_TWEEDLEDUM, "Tweedledum is required for these tests.")
 @ddt
 class TestBooleanExpression(QiskitTestCase):
     # pylint: disable=possibly-used-before-assignment
@@ -70,21 +74,8 @@ class TestBooleanExpression(QiskitTestCase):
 
         self.assertEqual(bool(int(result)), expected)
 
-    def test_errors(self):
-        """Tests correct identification of errors"""
-        exp = "x1 & x2 | x3"
-        with self.assertRaisesRegex(ValueError, "var_order missing.*x2"):
-            BooleanExpression(exp, var_order=["x1", "x3"])
 
-        bool_exp = BooleanExpression(exp)
-        with self.assertRaisesRegex(ValueError, "bitstring length differs.*2 != 3"):
-            bool_exp.simulate("01")
-        with self.assertRaisesRegex(ValueError, "bitstring must be composed of 0 and 1 only"):
-            bool_exp.simulate("012")
-        with self.assertRaisesRegex(ValueError, "'circuit_type' must be either 'bit' or 'phase'"):
-            bool_exp.synth(circuit_type="z_flip")
-
-
+@unittest.skipUnless(HAS_TWEEDLEDUM, "Tweedledum is required for these tests.")
 class TestBooleanExpressionDIMACS(QiskitTestCase):
     """Loading from a cnf file"""
 
@@ -100,32 +91,14 @@ class TestBooleanExpressionDIMACS(QiskitTestCase):
         self.assertEqual(simple.name, "simple_v3_c2.cnf")
         self.assertEqual(simple.num_qubits, 4)
         self.assertTrue(simple.simulate("101"))
-        self.assertFalse(simple.simulate("001"))
 
     def test_quinn(self):
         """Loads quinn.cnf and simulate"""
         filename = self.normalize_filenames("dimacs/quinn.cnf")
         simple = BooleanExpression.from_dimacs_file(filename)
         self.assertEqual(simple.name, "quinn.cnf")
-        self.assertEqual(simple.num_qubits, 17)
+        self.assertEqual(simple.num_qubits, 16)
         self.assertFalse(simple.simulate("1010101010101010"))
-
-    def test_bad_formatting(self):
-        """Tests DIMACS parsing on edge cases"""
-        # pylint: disable=trailing-whitespace
-        dimacs = """ 
-p cnf 10 5"""  # first line is not p cnf nor empty nor comment (it has whitespace)
-        with self.assertRaisesRegex(ValueError, "First line must start with 'p cnf'"):
-            exp = BooleanExpression.from_dimacs(dimacs)
-        dimacs = """p cnf 2 1
-         
-        1 2 0"""  # has empty line with whitespace - should ignore it
-        exp = BooleanExpression.from_dimacs(dimacs)
-        self.assertEqual(exp.name, "(x1 | x2)")
-
-        bad_filename = "bad_filename"
-        with self.assertRaisesRegex(FileNotFoundError, f"{bad_filename} does not exist"):
-            BooleanExpression.from_dimacs_file(bad_filename)
 
 
 if __name__ == "__main__":
