+++ test/python/synthesis/dimacs/simple_v3_c2.cnf
@@ -1,0 +1,5 @@
+c  simple_v3_c2.cnf
+c
+p cnf 3 2
+1 -3 0
+2 3 -1 0
